backend/web/_test/openapi/requirements.txt
astroid==2.5.3
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
isort==5.8.0
jedi==0.18.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.7.0
openapi-core==0.13.7
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
parse==1.19.0
parso==0.8.2
pylint==2.7.4
pyrsistent==0.17.3
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
